backend: added middleware for admin protection

+++ covidVaccination-laravel-backend/routes/api.php
@@ -25,12 +25,11 @@ Route::group(['middleware' => ['api', 'auth.jwt']], function () {
     Route::post('auth/logout', [\App\Http\Controllers\AuthController::class, 'logout']);
 
     // User
-    Route::post('user', [\App\Http\Controllers\UserController::class, 'save']);
-    Route::put('user/{ssn}', [\App\Http\Controllers\UserController::class, 'update']);
-    Route::delete('user/{ssn}', [\App\Http\Controllers\UserController::class, 'delete']);
-    Route::get('users', [\App\Http\Controllers\UserController::class, 'getAllUsers']);
+    Route::put('user/{ssn}', [\App\Http\Controllers\UserController::class, 'update'])->middleware('role:admin');
+    Route::delete('user/{ssn}', [\App\Http\Controllers\UserController::class, 'delete'])->middleware('role:admin');
     Route::get('user/{ssn}', [\App\Http\Controllers\UserController::class, 'findBySSN']);
     Route::get('user/isAdmin/{ssn}', [\App\Http\Controllers\UserController::class, 'isAdmin']);
+    Route::get('users', [\App\Http\Controllers\UserController::class, 'getAllUsers'])->middleware('role:admin');
 
     // Reservations
     Route::post('reservation', [\App\Http\Controllers\ReservationController::class, 'save']);
@@ -40,22 +39,25 @@ Route::group(['middleware' => ['api', 'auth.jwt']], function () {
     Route::get('reservations/vaccination/{vaccination_id}', [\App\Http\Controllers\ReservationController::class, 'getAllOfVaccination']);
 
     // Vaccinations
-    Route::post('vaccination', [\App\Http\Controllers\VaccinationController::class, 'save']);
-    Route::put('vaccination/{id}', [\App\Http\Controllers\VaccinationController::class, 'update']);
-    Route::delete('vaccination/{id}', [\App\Http\Controllers\VaccinationController::class, 'delete']);
+    Route::post('vaccination', [\App\Http\Controllers\VaccinationController::class, 'save'])->middleware('role:admin');
+    Route::put('vaccination/{id}', [\App\Http\Controllers\VaccinationController::class, 'update'])->middleware('role:admin');
+    Route::delete('vaccination/{id}', [\App\Http\Controllers\VaccinationController::class, 'delete'])->middleware('role:admin');
     Route::get('vaccinations2', [\App\Http\Controllers\VaccinationController::class, 'getAllRaw']);
     Route::get('vaccination/{id}/users', [\App\Http\Controllers\VaccinationController::class, 'getUsersOfVaccination']);
 
     // Locations
-    Route::post('location', [\App\Http\Controllers\VaccinationLocationController::class, 'save']);
-    Route::put('location/{id}', [\App\Http\Controllers\VaccinationLocationController::class, 'update']);
-    Route::delete('location/{id}', [\App\Http\Controllers\VaccinationLocationController::class, 'delete']);
+    Route::post('location', [\App\Http\Controllers\VaccinationLocationController::class, 'save'])->middleware('role:admin');
+    Route::put('location/{id}', [\App\Http\Controllers\VaccinationLocationController::class, 'update'])->middleware('role:admin');
+    Route::delete('location/{id}', [\App\Http\Controllers\VaccinationLocationController::class, 'delete'])->middleware('role:admin');
 
     // States
-    Route::post('state', [\App\Http\Controllers\StateController::class, 'save']);
-    Route::put('state/{state}', [\App\Http\Controllers\StateController::class, 'update']);
-    Route::delete('state/{state}', [\App\Http\Controllers\StateController::class, 'delete']);
+    Route::post('state', [\App\Http\Controllers\StateController::class, 'save'])->middleware('role:admin');
+    Route::put('state/{state}', [\App\Http\Controllers\StateController::class, 'update'])->middleware('role:admin');
+    Route::delete('state/{state}', [\App\Http\Controllers\StateController::class, 'delete'])->middleware('role:admin');
 });
+
+// Users
+Route::post('user', [\App\Http\Controllers\UserController::class, 'save']); // registration
 
 // Vaccinations
 Route::get('vaccinations', [\App\Http\Controllers\VaccinationController::class, 'getAllVaccinations']);
